requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
sqlalchemy==2.0.30
pydantic==1.10.14
python-dotenv==1.0.1
python-multipart==0.0.9
aiosqlite>=0.19.0

# ✅ Password hashing libraries
passlib[bcrypt]==1.7.4
bcrypt==3.2.2  # ✅ stable version compatible with passlib
